Introduce a more nuanced job retrying process.
This allows executors to raise a special type exception
(RetryableException) with an optional reschedule_for argument.

If reschedule_for is set the job will be moved to "queued" again, and
the execute_after set to that datetime.

If it is not set, the job status will be set to "exception" and the job
will remain in the jobs table.

If any other exception occurs the job existing default behaviour applies
(i.e. remove from jobs table, move to stats table with "exception" status)

Also introduce "attempts" to the jobs table. Migration to come.